refactor(search applications): rename engine view file
